Update bli_type_defs.h
Set BLIS_NUM_ARCHS automatically by inserting it as the last member in the enum. Also remove `=0` on first enum member since it is guaranteed by the standard. This make is easier and less error-prone to reorder enum entries.

This change was also included in changes I made to #344 and #345 in remerging them, but I thought it important to make this change earlier rather than later.