Fix bug in storing microkernel result
Details:
There was a bug in the DGEMM P10 kernel that did not store the result of the microkernel correctly. The behavior of the bug was that when the result was being stored. The reason was due to the wrong index calculation :(. This bug occured when I reorganize the kernels/power10/3 directory.