Add two new sandboxes.
The 'gemmlike-tci' sandbox is a copy of the 'gemmlike' sandbox, except that thread management has been switched to TCI (github.com/devinamatthews/tci). Additionally, macro definitiones for the typed framework code has been avoided in favor of a type-erasure approach.

The 'gemmlike-tci-cxx' sandbox further incorporates modern C++ features and takes advatntage of the TCI C++11 interface for a much cleaner code. This version does not currently build due to deficiencies in the Makefile.