Add back explicit checks for her2k/syr2k/herk/syrk.
NOTE: previously these checks happened in the _front layer, where cntx is guaranteed to be valid. Now the checks happen earlier when cntx may be NULL. This causes a segfault in bli_check_sufficient_stack_buf_size. Because the stack size will be checked again later, we now return success in the stack size check when cntx == NULL.