Move pba, sba pool, and mem_t to the thrinfo_t object.
Moving the sba pool and pba allows the rntm_t to be effectively const (although it is sometimes copied internally and modified to reflect different ways of parallelism). Moving the mem_t sets the stage for sharing a global control tree amongst all threads. Note that trsm threading is broken and is fixed in a later commit.