Update pack schema determination.
Don't store the *intended* packing schema in the objects, only the *actual* schema once the object is packed for real. This was necessary before to pass the schemas down to the thread decorator which then used that information to create the control tree. Since control tree creation is in bli_*_ex now this is no longer necessary. The schemas also have to be separately passed to bli_gemm_md which is kind of ugly, but that code is ripe for refactoring anyways.