Start putting more information directly in the control tree.
In this phase, add the threading granularity (block size multiple) and weighted threading preference. These are now determined and set in the control tree initialization (gemm and trsm), and stored in a part_cntl_t struct. This information is now read in the gemm/trsm variants and passed directly to bli_thread_range_[mn]dim. Note that this change requires passing the matrix objects to the control tree intitialization (along with the cntx_t). For now, the schemas are determined separated and also passed in---in the future the schemas will also be determined within the control tree initialziation so that these functions have signatures much like bli_gemm_ex itself (except the rntm_t).