Encode gemm/gemmtrsm microkernel and row-preference in the control tree.
This entailed making new control tree nodes for the gemm and trsm macro-kernels (gemmt and trmm use the gemm one). Since it is expected that there is a thrinfo_t node for BOTH the jr and ir levels of parallelism, a "dummy" empty control tree node is attached to the macrokernel node.