add runtime selection of power architecture
use "./configure power" and "make" to create a single shared library for
power10/power9/generic.
With this feature, blis shared library could be compiled on a power9 system
and run on a power10 system and vice-versa. If run on a generic system,
unoptimised generic code will execute.

This will work with gcc only, as that is the only compiler supported by both
power9 and power10 configurations in BLIS