Auto-detect the RISC-V ABI of the compiler and use -mabi= during RISC-V builds
Generate a build error if there is a 32/64-bit mismatch between the RISC-V ABI
or architecture and the BLIS configuration selected.

Handle Q, Zicsr, ZiFencei, Zba, Zbb, Zbc, Zbs and Zfh extensions in the RISC-V
architecture auto-detection. ZiFencei and Zicsr is not detectable with built-in
RISC-V macros right now.

ZiFencei is not important for BLIS because doesn't it have Just-In-Time
compilation or self-modifying code, and Zicsr is implied by the floating-point
extensions, which are required for good performance in BLIS.