add complex return detection for nvfortran
historically, PGI->NVHPC Fortran has used the Intel convention.
this is true on x86_64 and ppc64le, and has been true on aarch64.
however, starting with the 23.9 release, on the aarch64 platform,
nvfortran will switch to the gfortran convention.

this patch adds the appropriate logic to configure.
this patch has not been tested at all.  i am relying on CI for that.

Signed-off-by: Jeff Hammond <jehammond@nvidia.com>