CMake: Removing blatest-related targets for Windows/shared libs.
Due to the way dlls resolve internal symbols, calling custom function
xerbla_() is not possible. Because of that the targets result in errors
which are independent of BLIS library.
Testing Windows/static version will suffice.

To enable make check and make test targets, we throw warnings for the
Windows/shared cases and not depend on checkblas.

AMD-Internal: [CPUPL-2748]
Change-Id: Iaa93399dec5781277ee94611074f5ed4e70bcb37