Fixed linking issue with bli_print_msg function
Description:

In recent changes bli_print_msg is used in lpgemm
test application file bench_lpgemm.c for printing error
message. bli_print_msg is a blis library function which
is not exported for the usage of applications, because
of which linking failed when blis shared library is used
to build.

Updated bli_print_msg with printf in the bench_lpgemm.c

AMD Internal: CPUPL-5326

Change-Id: I021849baa6881bd997013e42013db1c5c711627f